refactor: setup user model in a separate folder, update imports and use method getByEmail directly from the model
